+++ changelog.md
@@ -21,6 +21,7 @@ Codion Change Log
 ### is.codion.swing.common.model
 - FilterTableModel.TableSelection moved to list package and renamed FilterListSelection, related refactoring.
 - FilterTableSort.sorted() moved to FilterListModel.Sort.
+- DefaultFilterTableItems.DefaultVisibleItems no longer adds a table model listener, changes to visible items notified explicitly.
 ### is.codion.swing.common.ui
 - LookAndFeelSelectionDialogBuilder.createControl() overloaded without parameter.
 - BackgroundColorValidIndicator bug fixed, did not refresh state when colors changed during updateUI().